+++ internal/err_handler_registry.go
@@ -1,4 +1,4 @@
-package core
+package internal
 
 import (
 	"errors"
